oroi: Create CSV downloading system
This adds a system to download either a full dump of the data or the data
from a particular query.

Full csv dumps are pre-generated from the relational database
using ./manage.py csv_user_dump The output locations and fields are set in the
settings. The suggested procedure is to run this after each load.

Query dumps are generated via the ElasticSearch database using the same
API as the UI uses.